my mistake is that the output was in float, I changed float to "int" in line 8, added int. to lines 13, 15, 17, and 19.

+++ change.py
@@ -5,18 +5,18 @@
 
 
 print("Please enter an amount in cents less than a dollar.")
-num_1=float(input())
+num_1=int(input())
 quarters=25
 dimes=10
 nickels=5
 pennies=1
-num_quarters=(num_1//quarters)
+num_quarters=int(num_1//quarters)
 cents_remq=(num_1%quarters)
-num_dimes=(cents_remq//dimes)
+num_dimes=int(cents_remq//dimes)
 cents_remd=(cents_remq%dimes)
-num_nickels=(cents_remd//nickels)
+num_nickels=int(cents_remd//nickels)
 cents_remn=(cents_remd%nickels)
-num_pennies=(cents_remn//pennies)
+num_pennies=int(cents_remn//pennies)
 print("Your change will be:")
 print("Q:", num_quarters)
 print("D:", num_dimes)
